+++ sshd-core/src/main/java/org/apache/sshd/common/io/IoWriteFuture.java
@@ -34,5 +34,4 @@ public interface IoWriteFuture extends SshFuture<IoWriteFuture>, VerifiableFutur
      * between the two.
      */
     Throwable getException();
-
 }